Expand reductions to HL ops instead of LL ops
Mostly for separation of concerns. It's a bit weird to have the reduction expansion be lowering in addition to expanding to fat ops.

For now this will cause a performance hit because of the extra copies, but we're more concerned with ergonomics atm. Eventually we'll have a pass to elide the unnecessary copies.

PiperOrigin-RevId: 270840295