Rationalize Interpreter scalar copies with validation to avoid runtime failures
Some of these passes generate copies that would fail in the interpreter kernel because it expects all ranks and index counts to be the same. Instead, add validation to the op and avoid mismatches between the ranks of src and dst and the sizes of the indices and lengths.

This allows cleaning up the treatment of scalar copies so they aren't special cased to pretend they're rank 1

PiperOrigin-RevId: 271999245