Remove misleading type conversion
Type conversion is not actually happening as part of this pass (the converter isn't even passed to the conversion framework), so this is currently a no-op. MemRef conversions happen as part of prior custom passes and without these passes the conversions here wouldn't work because they don't cover the FuncOp itself, which must be explicitly converted.

We can move this conversion over to the conversion framework, but for now it's misleading to have the code for it here.

PiperOrigin-RevId: 275434055