Always log to stderr instead of stdout
This seems like a much more reasonable default. We could later set up the option for log files, if that's desirable and just using standard unix redirects for stderr isn't sufficient.

Also fixes the e2e tests that are failing because FileCheck looks at the stdout of iree-run-mlir and gets confused by the log lines. Closes #96

PiperOrigin-RevId: 276703432