Check that global invocation ID is within the specified workload range.
Guard the generated SPIR-V kernel such that only those workitems
within the range of the workload specification execute the body of the
function. This allows setting an arbitrary workgroup size, instead of
the current 1x1x1.

Add a new attribute iree.executable.workgroupsize on dispatch
functions that sets the workgroup size to be used for the SPIR-V
modules. It is currently hard-wired to 32x1x1.

PiperOrigin-RevId: 277966689