Capture default FileCheck args in a wrapper script
This way we don't have to declare them every time.

One disadvantage here is that this prevents us from redeclaring these arguments because FileCheck bails if an argument is declared multiple times. That means removing the option to do something like "--dump-input=always". We could make this more sophisticated at some point.

Note that this does not update the TF interop tests that use FileCheck because they're doing some additional magic that means it can't locate the FileCheck binary. Probably this is surmountable, but would require some more investigation.

PiperOrigin-RevId: 283092455