Use checkout action in workflows (#256)
This should simplify things a bit and will hopefully be more robust so we get fewer errors during checkout.

Tested: presubmit (clang-format) and cmake actions still pass. Bazel action is suuuuper slow, so I'm not going to wait for it, but it at least gets to the point of starting the build.