Initialize stab at goo for wrapping C++ native modules and dealing with the VM ABI. This removes almost all of the boilerplate for defining new modules and makes it super easy to model tuples, fixed arrays, and variadic argument types in-to/out-of of the VM. Future improvements around type registration and C++ RefObject interop are planned.
As verified here there is nearly zero overhead (vs. the manual approach) and in many cases less than when hand-written: https://godbolt.org/z/6CX3n5
For functions taking integers the entire trampoline is on the order of ~5 instructions. Ref object usage is all move based and there are no additional retains/releases (beyond what is required for balancing the ref count). There's the compile-time overhead of the templates but they are relatively simple and always isolated to a single compilation unit (no global cross-unit explosion).

PiperOrigin-RevId: 288044390