Restructure kokoro build configuration directory
This allows us to separate the bazel and cmake build and share configuration between jobs in each. The configuration sharing also happens in the internal job configurations which indicate how to trigger the job and talk to the GitHub API, where the sharing is more than the single line here.

I've copied the bazel configurations and will delete the old ones once everything is moved over to avoid disruption in the currently running jobs. The new jobs will post statuses as "linux_bazel" rather than the generic "kokoro"

PiperOrigin-RevId: 290150545