Make bazel lit test suites less magic
Now that we're using the same lit runner internally and in OSS, we can get rid of a lot of the magic here. This obviates the need for a separate setup and glob rule and makes the dependencies of each target explicit, following bzl best practices. It should also make this easier to automatically translate to cmake, which I'll work on separately.

Note that this turns the recursive globs into non-recursive globs. This does not cause us to exclude tests, and in fact stops running tests under iree/compiler/Translation/Interpreter/Transforms/test/xla twice, which we were doing before. To verify, `cat ~/tmp/open_files.txt | grep "/BUILD" | sed -e 's@/BUILD@@' | xargs -I{} find {}/ -type d | grep -v "/$"`, which gives all the subdirectories of the directories containing the modified BUILD files.

PiperOrigin-RevId: 291435291