Pass kwargs to the test_suite in iree_lit_test_suite
It's reasonable to expect that any kwargs added here would be forwarded to the test_suite if applicable. This introduces the possibility of someone passing a kwarg that only applies to the tests themselves, but the failure mode here is an immediate error rather than silently ignoring the kwarg as in the current implementation. It's also an easy fix of just adding it as an explicit named argument. Rather than list all possible named arguments of sh_test that aren't applicable to test_suite, we'll leave that for when they're needed.

These additional kwargs that only apply to sh_test are: deps, args, bash_version, exec_compatible_with, exec_properties, flaky, local, shard_count, timeout, toolchains

PiperOrigin-RevId: 293239207