Add IREE `tensorlist` dialect.
This is the beginning of TF TensorList/TensorArray support in IREE.
Right now, the strategy is to be as braindead about the lowering as
possible, to avoid getting tripped up by any nasty TF behavior.

Once we have e2e correctness, we can evaluate alternative approaches.

This is part of a stack of CL's. The others are:
- `tensorlist` IREE VM native module
- `tf_tensorlist` dialect
- Python e2e plumbing glue.

PiperOrigin-RevId: 295221980