Set resource subtype correctly on bound inputs.
This gives more information to shape propagation.

To do this, we need to insert a cast back to the raw imported type, but
shape propagation will eliminate that later.

Incidentally, this also standardizes the arg types used to represent bound
inputs between the V1 (SignatureDef) and V2 (ObjectGraph) importers. As
such, I've fixed a TODO in the verifier to actually verify that the args
have the right type.

PiperOrigin-RevId: 296009270