Properly broadcast intermediate shape for vmla broadcast_in_dim and gather.
The original transforms were all static shape based and included a general reshape to get inputs ready to tile/copy. In the dynamic shape world, this is trickier and necessitated building a degenerate broadcast sequence of ops to dynamically extend the shape.

With this, all test/e2e/xla tests pass when forced to use VMLA.

Fixes #940 (and also broadast_in_dim)

PiperOrigin-RevId: 299221389