Convert backends to handle dispatch regions with reduce ops.
With -iree-flow-experimental-dispatch-reduce, reduction dispatch
functions are just dispatch functions with a single reduction op in
them. Adapt the HLOToReduction pass to handle this (instead of the
reduction dispatch region).
This change also removes the now defunct reduction code-generation
within the XLAToSPIRV conversion, i.e. index computation path.
With this change, also enable some of the e2e tests that were
dependent on reduction with initialization to work.

PiperOrigin-RevId: 299879419