Switch abs test to check framework
This encompasses a lot of weirdness on different backend support and includes updating the build rules to allow this. I opted for a new function for the single backed/driver case even though you could just pass a one-element list to iree_check_test_suite because I wanted to avoid the case where the generated names become super verbose as the same information is added multiple times.

PiperOrigin-RevId: 305913199