Adds Dockerfiles for building with bazel and updates GitHub CI workflows
- Adds Dockerfiles that define images for our bazel and bazel with tensorflow builds.
- Adds a script for updating these images on the IREE-OSS Google container registry (GCR).
  - These images are hosted as `gcr.io/iree-oss/bazel` and `gcr.io/iree-oss/bazel-tensorflow`
- Updates the GitHub CI workflows to use these images.

This change adds Dockerfiles specifying the dependencies of IREE's core bazel build and IREE's integrations/bindings builds. This allows us to run CI tests in GitHub Actions without having to install the dependencies for each commit to master.

We are hosting the Docker images that these files build on the IREE-OSS GCP project. Currently, this means that every time there is a change to one of the Docker images, someone who is a storage admin for IREE-OSS must run the `build_tools/docker/update_gcr_images.sh` script. It should be possible to automate this process whenever a change is made to these files on GitHub in a future PR.

This change was tested by creating another branch based off of this one, changing the workflows to run on push to that branch and then pushing that change. Links to each test:

- [Bazel Build - Bindings](https://github.com/meadowlark@google.com/iree/actions/runs/100891735)
- [Bazel Build - Fallthrough](https://github.com/meadowlark@google.com/iree/actions/runs/100891736)
- [Bazel Build - Integrations](https://github.com/meadowlark@google.com/iree/actions/runs/100891738) (Failed due to the build being broken at the time of testing).
- [Bazel Build -Core](https://github.com/meadowlark@google.com/iree/actions/runs/100891741)

Since the integrations build failed due to the state of the repo at the time it was run, I ran the test again on my machine on a commit with a working build as a stand in and it passed.

Closes https://github.com/google/iree/pull/1874

COPYBARA_INTEGRATE_REVIEW=https://github.com/google/iree/pull/1874 from meadowlark@google.com:docker-containers b280481b306a1a2ba76758437930a235ef6cf4f7
PiperOrigin-RevId: 311007783