Handle i1 types on ExtractElementOp before lowering to IREE::Flow::TensorLoadOp.
It's impossible to allocate a buffer of element size i1. The change converts i1
types to i8 types, so we can allocate and load from the buffer. This will
benefit all the backends, so we won't need to handle i1 storage at codegen side.

This change doesn't address the issue generally, a more robust and general i1
support is needed (part of https://github.com/google/iree/issues/2050)

PiperOrigin-RevId: 316565087