Fixes a bug where Kokoro's artifacts directory wasn't being deleted. (#2434)
This should fix the issue where our CMake presubmits would sometimes take twice as long as they should.

Explanation:
The problem was that (for some reason) `bash` was evaluating `rm -rf "${KOKORO_ARTIFACTS_DIR?}/*"` to `rm -rf '/tmpfs/src/*'`. The additional single quotes would cause `rm` to fail with a directory not found error, which would be swallowed by the `-f` flag.

Removing the double quotes around the `/*` expansion removes the single quotes that appear when `bash` + `rm` evaluates it. This gives a permission denied error, which we fix by adding `sudo`.