Stop always sending TF update PRs to gcmn (#2634)
This sends me a lot of email and doesn't add much value, I think.

There is the option to use team reviewers. We could look at creating a
team for the build cops. It would be really nice to request review from
or tag someone dynamically based on our build cop rotation, but not sure
how to do that.