[vulkan] Reset TimePointFence status when releasing back to pool (#2905)
This obviously is required to make sure that when later we acquire
the same TimePointFence, it's not in signaled state to start.

This bug is revealed by a non-deterministic crash on SwiftShader
where a command buffer can be destroyed in the middle of being
executed. This is because in IREE after queue submission we wait
on the TimePointSemaphore to signal to a specific value. Under
the hood, TimePointFence is used to query the status from GPU.
But it's always signaled and it causes IREE side to progress and
delete the command buffer (given it's already submitted and
executed seen by IREE). This triggers command buffer deletion
in SwiftShader, while it's still executing it.