Rename main source files used directly in cc binaries (#3249)
These naming diffs are frequently confusing, I find. This consistently
names the file after the primary binary in which it is used (if
applicable) suffixed with `main.cc` separated with a dash or an
underscore depending on what is used in that binary.

This is also motivated by absl flags behavior. Absl uses the name of
the binary to determine which flags to include in `--help` by default:
those from source files whose basename match the binary name with an
optional (`-main` or `_main` suffix) get included.

We do have the option of customizing the absl behavior instead if we 
have strong feelings about alternative naming schemes.

Part of https://github.com/google/iree/issues/3247