Support PadOp on Linalg on tensors. (#4983)
linalg.pad_tensor cannot be represented as a structured
operation. Instead it is lowered to

%b = linalg.fill(%dest, %pad_val)
%c = subtensor_insert %source into %b[%low_pad_0, %low_pad_1, ...]

This requires handling subtensor_insert when converting from tensors
to buffers. This can be done effectively only when it is possible to
do in-place updates within dispatch regions. So currently only support
this on CPU side.