Adapt pooling ops lowering to Linalg on tensors. (#5024)
This also introduces the concept of bufferization of shape-only operand.
If an init_tensor op can not map to a buffer, a fake/temp memref will be
allocated. And we rely on later passes to delete the alloc op. This is
similar to what we did to lower mhlo.reduce_window to linalg.pooling_*.