Removing a sketchy check that was preventing releases of ref ptrs. (#5258)
When moving the target must be released regardless of whether the pointers
match in order to keep things balanced. This could happen in cases where
a register contained a ref ptr and a function call returned the same
pointer with the move bit set, leaking the release and failing to clobber.
Fixes #5141.