Adding command buffer queue affinity. (#5265)
This is needed for the CPU backend and vulkan tracing. It's a somewhat
unfortunate dependency issue and ideally we wouldn't need to track the
queue during recording however for one shot recordings we know at
compile-time what queue a stream will be submitted to. We can always
change the behavior to ignore/treat as suggestion the affinity when
recording reusable command buffers.