Adding a minimal flag parser. It does most of what we want without any C++ in 3 modes: - entirely disabled (turned into some static variables) - argc/argv parser only for main() - flagfile parsing that uses fopen/fread
It also has the nifty property of a round-tripping mode that lets you
get all the current flag values and pipe them to a file to then load
back in a flagfile, which should make it easier to reproduce
flag-dependent configurations.

Likely lots of little tweaks remaining (there always are), but this may be
enough to start replacing our existing usage of the abseil flags. One
thing mostly implemented is callbacks so that we can support lists
(multiple instances of a flag on the command line) and actions, which
aren't great in abseil flags.

Fixes #3814.