Adding `-iree-vm-emit-polyglot-zip` flag. (#5709)
This allows .vmfb outputs to be opened/unzipped as if they were .zip files.
The behavior isn't required by the runtime and this is just to make getting
binary blobs out of the flatbuffers easier. Specifically, it's now possible
to extract the executable binaries in ELF format. SPIR-V and PTX still
have flatbuffer wrappers so they aren't directly usable yet but that will
be fixed in the future.