[spirv] Migrate ConvertToGPUPass' invocation tiling logic (#5814)
`ConvertToGPUPass` is a sink for lowering away all Linalg ops
in the SPIR-V pipeline: it can distribute a Linalg op to both
global invocation IDs (if just having one-level distribution) or
local invocation IDs (if having two-level distribution). 

This is mostly historical; now we have more proper layering
in the pipeline where we perform the first-level tiling and
distribution at flow level and the second/third level at
`TileAndVectorizeInOneWorkgroupPass`, the functionality
in `ConvertToGPUPass` is overlapping with that, although
in a complementary way: the tiling in the former pass does
not handle the cases where we cannot do imperfect tiling;
it requires perfectly tiled cases, as we are assuming number
of processors equal to number of iterations to avoid generating
`scf.for` loop from the start. `ConvertToGPUPass` is more
generic and can handle all cases.  This is all opaque and
complicated. 

This commit relaxes the `TileAndVectorizeInOneWorkgroupPass`
to not assume the number of processors equal to the number
of iterations. Now we just tile and cyclically distribute using
`scf.for` loops. This causes issues for perfectly tiled cases
as we need to canonicalize the `affine.min` and one-trip
`scf.for` away to expose static sizes for further vectorization.
That can be done by pulling in additional canonicalization
patterns.

Also in order to utilize `TileAndVectorizeInOneWorkgroupPass`
for the second/third level tiling we need to have the corresponding
scheme in launch configuration for them. This commit also
adds the default second/third level tiling for all now supported
Linalg ops: no tiling on subgroup and tiling to 1 for invocations.
This at the same time helps to clean up a bunch of unwieldy
templated configurations for different ops..

Together, the above migrates the invocation tiling logic
in `ConvertToGPUPass` to their proper places. This is the
first step as reining in `ConvertToGPUPass` and launch
configurations.