Handle removing unused results of a CloseureInterfaceOp (#5969)
The canonicalization pattern to remove a result produced by the
ClosureInterfaceOp needs to make sure that all uses within the
dispatch region are write-only and that these operations are erased
before dropping the arguments.
Add an interface method to check if an argument to the
ClosureInterfaceOp region has any read uses. This is used to determine
if the result/operand can be dropped.