Drop usage of LLVMCodegenOptions. (#6084)
The use of LLVM style command line flags makes it harder to print out
reproducers or have a proper string representation of the pass
pipeline. It is better to use MLIR wrappers around these options. With
dynamic pass pipelines, there is a point to anchor these options.

Some of the options are now moved into specific passes. like the
unfusedFMAOps option is part of the ConvertToLLVM pass.

Some unnecessary tests are deleted.

Also enable use of vectorization pipeline on the VMVX backend. Since
the VMVX backend cannot handle vectors or vector operations, add a
flag to avoid vectorization but still do the tiling for the different
cache levels.

Fixes #5925