Revisit scf.for bufferization. (#6125)
A few minor adjustments to scf.for bufferization.

When tracking number of uses, specifically look for destructive
update patterns to treat than as "non real" uses since these uses
are due to tensor abstraction limitations.
Refactor the walk to store and also stop the walk at the
subtensor_insert that is a destructive update.
During scf.for conversion, replace the operands of the scf.yield
with the values for the arg that represent the initArg of the
loop. This effectively marks the (result, initArg, regionArg) set as
a no-op that is canonicalized later.
Remove the now unecessary loop manipulation added previously to make
the loop dead.