Handle read-only tensors as `initArgs` of `scf.for`. (#6212)
When a read-only tensor is passed as an initArg to an scf.for, it
cannot be tied to the result to do in-place updates. Plumbing this
through prompted some more cleanups.

Make the destructive update pattern recognizer more strict, and to
have separation of concerns, do a separate walk to find destructive
updates and tie operands accordingly.
Make the walk of use-def chains (which walks uses), utilize the
BufferizationPlan to figure out which result uses to walk next.
Also add support for scf.if, tensor.insert operations to
bufferization pass.