Remove dependency on llvm-bazel (#6298)
Instead we use the Bazel build files now upstream in the LLVM
repository. There is a risk here because we can't go back and fix these
build files and continue to point at the same commit, so if the LLVM
integration into google3 uses a commit with broken build files, we won't
be able to update for a bit. I think the real solution to that may be to
create our own fork of LLVM where we can happily cherry-pick.