Add 15 minute test timeout to all ctest invocations (#6617)
Hanging forever is extremely unhelpful. Setting test-specific timeouts,
which could be derived from the Bazel timeouts as part of Bazel->CMake
would be helpful, but in the meantime, just setting a reasonable upper
bound should get us a lot of mileage. This matches the Bazel timeout for
"long" tests, which is the maximum timeout we have for all of our tests
that we run in our CI.

Part of https://github.com/google/iree/issues/5121