[Cleanup] Delete unused flags. (#7365)
The matmulVectorSize flag is actually unsued in LLVM CPU codegen because
the native_vector_size attribute will always be set in entry point
function. However, we can not expect every entry point function has the
attribute. Because VMVX backend uses the config and does not have the
attribute. In this case, the PR merges matmulVectorSize into
clNativeVectorSizeInBytes.

Also fixes a typo in error message.