Use `PartitionableLoopInterface` on the compilation stack. (#8197)
This PR uses the `PartitionableLoopInterface` to know which loops to distribute in a manner consistent between Flow Dialect and Codegen. 

Use of this interface know allows not distributing unit-dimensions, that generates better code for that corner case. In the process some clean up of the configuration selection (on the CPU side) to get more ops going through the vectorized path.

Co-authored-by: Lei Zhang <antiagainst@google.com>